# Use AtomicInteger + mod instead of synchronizing threads in CountingS…